+++ Eligible.NET/src/EligibleService.Model/PaymentStatus/Amount.cs
@@ -5,7 +5,7 @@ namespace EligibleService.Model.PaymentStatus
     public class Amount
     {
         [JsonProperty("billed")]
-        public int? Billed { get; set; }
+        public double? Billed { get; set; }
 
         [JsonProperty("paid")]
         public double? Paid { get; set; }
